requirements.txt
django-crontab
celery
django-celery-beat
requests